style: `enum` cannot be used with import type
