+++ packages/elements/src/overlay-menu/helpers/constants.ts
@@ -1,0 +1,7 @@
+export enum Navigation {
+  FIRST = 'First',
+  LAST = 'Last',
+  NEXT = 'Next',
+  PREVIOUS = 'Previous',
+}
+
