+++ packages/elements/src/overlay-menu/helpers/types.ts
@@ -8,11 +8,3 @@ export type NestedMenu = {
 };
 
 export type OverlayMenuData = ItemData[] | CollectionComposer<ItemData>;
-
-export enum Navigation {
-  FIRST = 'First',
-  LAST = 'Last',
-  NEXT = 'Next',
-  PREVIOUS = 'Previous',
-}
-
